refactoring: move PresburgerSet into analysis
