+++ mlir/lib/Dialect/Presburger/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_mlir_dialect_library(MLIRPresburgerDialect
   PresburgerDialect.cpp
   PresburgerOps.cpp
-  Set.cpp
   Attributes.cpp
   Types.cpp
   Parser.cpp
